[Chore]MsCommandをMineStackCommandに移植
