+++ src/main/scala/com/github/unchama/seichiassist/commands/MsCommand.scala
@@ -42,25 +42,7 @@ object MsCommand {
         }
       )
     )
-    .execution { context =>
-      val args = context.args.parsed
-      val categories: Map[Int, MineStackObjectCategory] = Map(
-        1 -> MineStackObjectCategory.ORES,
-        2 -> MineStackObjectCategory.MOB_DROP,
-        3 -> MineStackObjectCategory.AGRICULTURAL,
-        4 -> MineStackObjectCategory.BUILDING,
-        5 -> MineStackObjectCategory.REDSTONE_AND_TRANSPORTATION,
-        6 -> MineStackObjectCategory.GACHA_PRIZES
-      )
-      val category = categories.get(args.head.toString.toInt)
-      category match {
-        case Some(_category) =>
-          val _page = args(1).toString.toInt - 1
-          IO.pure(CategorizedMineStackMenu(_category, _page).open)
-        case None =>
-          IO(MessageEffect("不明なカテゴリです。"))
-      }
-    }
+    .execution { context => }
     .build()
     .asNonBlockingTabExecutor()
